Tribeca integration (#1)
* Add tribeca SDK

* fix x padding on smaller screens

* gov onboarding

* WIP governor layout

* Link to proposal creator

* proposal WIP

* draft

* proposal link

* proposal rendering

* profile

* 1024

* mb 20

* header width

* show numbers

* more format

* voting wip

* add lockup mechanism

* proposal activate

* voting power

* vote side

* voting fmt

* proposals list

* date format

* voting balance

* sot proposal events

* create proposal button

* Update deps

* package bumps

* Fix types

* fix yarn doctor output

* fix lint

* governance params

* wallet

* message diferent place

* More proposal actions

* Upgrade to Webpack 5, tailwind 3

* fix empty tx

* Update lockup

* updates

* coma

* Queue proposal

* update tailwind thing

* Fix gov spacing

+++ src/hooks/useParsedTX.ts
@@ -33,7 +33,7 @@ export const useParsedTX = (smartWalletKey: PublicKey, index: number) => {
         txData.accountInfo.data.instructions
           .map((rawIx) => ({
             ...rawIx,
-            data: Buffer.from(rawIx.data as Uint8Array),
+            data: Buffer.from(rawIx.data),
           }))
           .map((ix): Omit<ParsedInstruction, "title"> => {
             const idl = idls.find((theIDL) =>
